+++ src/api.ts
@@ -109,8 +109,8 @@ export class Api {
 		res.json({});
 	}
 
-	public async deleteDevice(req: express.Request, res: express.Response) {
-		log.info("Received delete device request");
+	public async proxyRequest(req: express.Request, res: express.Response) {
+		log.info(`Proxying request ${req.path}...`);
 		if (!req.session) {
 			this.sendStatus(res, STATUS_BAD_REQUEST, "M_UNKNOWN", "No session");
 			return;
@@ -121,49 +121,13 @@ export class Api {
 		}
 		try {
 			const hsRes = await got({
-				method: "DELETE",
-				url: `${this.homeserverConfig.url}/_matrix/client/r0/devices/${encodeURIComponent(req.params.deviceId)}`,
-				headers: {
-					Authorization: `Bearer ${req.accessToken}`,
-				},
-				json: { auth: {
-					type: "com.famedly.login.token",
-					identifier: {
-						type: "m.id.user",
-						user: req.session.data.username,
-					},
-					user: req.session.data.username,
-					token: this.generateToken(req.session.data.username),
-				}},
-			}).json();
-			log.info("Successfully deleted device!");
-			res.json(hsRes);
-		} catch (err) {
-			log.error("Couldn't reach matrix server!", (err.response && err.response.body) || err);
-			this.sendStatus(res, STATUS_INTERNAL_SERVER_ERROR, "M_UNKNOWN", "Backend unreachable");
-			return;
-		}
-	}
-
-	public async deleteDevices(req: express.Request, res: express.Response) {
-		log.info("Received delete devices request");
-		if (!req.session) {
-			this.sendStatus(res, STATUS_BAD_REQUEST, "M_UNKNOWN", "No session");
-			return;
-		}
-		if (!req.session.data.username) {
-			this.sendStatus(res, STATUS_BAD_REQUEST, "M_UNKNOWN", "No username/password found or bad password provider");
-			return;
-		}
-		try {
-			const hsRes = await got({
-				method: "POST",
-				url: this.homeserverConfig.url + "/_matrix/client/r0/delete_devices",
+				method: req.method as "GET" | "POST" | "PUT" | "DELETE" | "PATCH",
+				url: this.homeserverConfig.url + req.path,
 				headers: {
 					Authorization: `Bearer ${req.accessToken}`,
 				},
 				json: {
-					devices: req.body.devices,
+					...req.body,
 					auth: {
 						type: "com.famedly.login.token",
 						identifier: {
@@ -175,7 +139,7 @@ export class Api {
 					},
 				},
 			}).json();
-			log.info("Successfully deleted devices!");
+			log.info("Successfully sent request to homeserver");
 			res.json(hsRes);
 		} catch (err) {
 			log.error("Couldn't reach matrix server!", (err.response && err.response.body) || err);