+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.4.0
+ - Add `default` stage handler config
+ - Proxy device signing key upload
+
 ## 0.3.0
  - Make UIA dynamic
  - Add `com.famedly.login.welcome_message` stage